+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add support for merge request resource label event endpoints
 * Add support for merge request and merge request note award emoji endpoints
 * Add support for `MergeRequests::remove`
+* Add support for `MergeRequests::addToMergeTrain`
 * Add support for container registry endpoints
 * Add support for `Environments::stopStale`
 * Add support for `last_activity_after` and `last_activity_before` in `Groups::projects`